refactor: Question 내부 컴포넌트 분리, 폴더 구조 수정
